save close deprecation
