+++ src/pyedb/grpc/edb.py
@@ -378,23 +378,6 @@ class Edb(EdbInit):
                     except:
                         self.logger.info(f"Failed to delete AEDT project-related file {file}.")
 
-    def _clean_variables(self):
-        """Initialize internal variables and perform garbage collection."""
-        self.grpc = True
-        self._materials = None
-        self._components = None
-        self._core_primitives = None
-        self._stackup = None
-        self._padstack = None
-        self._siwave = None
-        self._hfss = None
-        self._nets = None
-        self._layout_instance = None
-        self._variables = None
-        self._active_cell = None
-        self._layout = None
-        self._configuration = None
-
     def _init_objects(self):
         self._components = Components(self)
         self._stackup = Stackup(self, self.layout.layer_collection)
@@ -1210,35 +1193,11 @@ class Edb(EdbInit):
             return PointData(x, y)
 
     @staticmethod
-    def _is_file_existing_and_released(filename) -> bool:
-        if os.path.exists(filename):
-            try:
-                os.rename(filename, filename + "_")
-                os.rename(filename + "_", filename)
-                return True
-            except OSError as e:
-                return False
-        else:
-            return False
-
-    @staticmethod
     def _is_file_existing(filename) -> bool:
         if os.path.exists(filename):
             return True
         else:
             return False
-
-    def _wait_for_file_release(self, timeout=30, file_to_release=None) -> bool:
-        if not file_to_release:
-            file_to_release = os.path.join(self.edbpath)
-        tstart = time.time()
-        while True:
-            if self._is_file_existing_and_released(file_to_release):
-                return True
-            elif time.time() - tstart > timeout:
-                return False
-            else:
-                time.sleep(0.250)
 
     def _wait_for_file_exists(self, timeout=30, file_to_release=None, wait_count=4):
         if not file_to_release:
@@ -1261,36 +1220,36 @@ class Edb(EdbInit):
     def close_edb(self) -> bool:
         """Close EDB and clean up resources.
 
+        ..deprecated:: 0.51.0
+           Use: func:`close` instead.
+
         Returns
         -------
         bool
             True if successful, False otherwise.
         """
-        self.close()
-        start_time = time.time()
-        self._wait_for_file_release()
-        elapsed_time = time.time() - start_time
-        self.logger.info("EDB file release time: {0:.2f}ms".format(elapsed_time * 1000.0))
-        self._clean_variables()
-        return True
+        warnings.warn("Use method close instead.", DeprecationWarning)
+        return self.close()
 
     def save_edb(self) -> bool:
         """Save current EDB database.
 
+        ..deprecated:: 0.51.0
+           Use: func:`save` instead.
+
         Returns
         -------
         bool
             True if successful, False otherwise.
         """
-        self.save()
-        start_time = time.time()
-        self._wait_for_file_release()
-        elapsed_time = time.time() - start_time
-        self.logger.info("EDB file save time: {0:.2f}ms".format(elapsed_time * 1000.0))
-        return True
+        warnings.warn("Use method save instead.", DeprecationWarning)
+        return self.save()
 
     def save_edb_as(self, fname) -> bool:
         """Save EDB database to new location.
+
+        ..deprecated:: 0.51.0
+           Use: func:`save_as` instead.
 
         Parameters
         ----------
@@ -1302,16 +1261,8 @@ class Edb(EdbInit):
         bool
             True if successful, False otherwise.
         """
-        self.save_as(fname)
-        start_time = time.time()
-        self._wait_for_file_release()
-        elapsed_time = time.time() - start_time
-        self.logger.info("EDB file save time: {0:.2f}ms".format(elapsed_time * 1000.0))
-        self.edbpath = self.directory
-        self.log_name = os.path.join(
-            os.path.dirname(fname), "pyedb_" + os.path.splitext(os.path.split(fname)[-1])[0] + ".log"
-        )
-        return True
+        warnings.warn("Use method save_as instead.", DeprecationWarning)
+        return self.save_as(fname)
 
     def execute(self, func):
         """Execute EDB utility command (Not implemented in gRPC).
